Change name for new class
