--- lib/move_checker.rb
@@ -1,4 +1,0 @@
-#frozen_string_literal: true
-
-class MoveChecker
-end
