+++ lib/move_examiner.rb
@@ -1,0 +1,7 @@
+#frozen_string_literal: true
+
+class MoveExaminer
+  def within_limits?(array)
+    array.all? { |num| num.between?(0, 7) }
+  end
+end
